+++ reggie/src/value/traits/return_representable.rs
@@ -7,7 +7,7 @@ use crate::{
 pub trait ReturnRepresentable {
     fn returns() -> FunctionSignatureList;
     fn to_lua_return(self, argument_registers: &mut ArgumentRegisters) -> Result<(), EvalError>;
-    fn return_count() -> usize;
+    fn return_count() -> u16;
 }
 
 impl ReturnRepresentable for () {
@@ -18,7 +18,7 @@ impl ReturnRepresentable for () {
         Ok(())
     }
 
-    fn return_count() -> usize {
+    fn return_count() -> u16 {
         0
     }
 }
@@ -31,7 +31,7 @@ impl ReturnRepresentable for LuaValue {
         argument_registers.d[0] = self;
         Ok(())
     }
-    fn return_count() -> usize {
+    fn return_count() -> u16 {
         1
     }
 }
@@ -45,7 +45,7 @@ impl ReturnRepresentable for (LuaValue, LuaValue) {
         argument_registers.d[1] = self.1;
         Ok(())
     }
-    fn return_count() -> usize {
+    fn return_count() -> u16 {
         2
     }
 }
@@ -60,7 +60,7 @@ where
     fn to_lua_return(self, argument_registers: &mut ArgumentRegisters) -> Result<(), EvalError> {
         self.and_then(|value| value.to_lua_return(argument_registers))
     }
-    fn return_count() -> usize {
+    fn return_count() -> u16 {
         T::return_count()
     }
 }
